requirement.txt
Django==2.1.4
django-crispy-forms==1.7.2
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
